+++ jsapp/js/components/AssetTagsModal/index.ts
@@ -1,0 +1,2 @@
+export * from './AssetTagsModal'
+export * from './openAssetTagsModal'
